% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSGPFlowMatrices.R
\name{addsgpFlowMatrices}
\alias{addsgpFlowMatrices}
\title{Add SGPFlow Matrices to SGPstateData}
\usage{
addsgpFlowMatrices(
  state,
  year,
  cohort_type,
  add.matrices.to.which.state = NULL
)
}
\arguments{
\item{state}{The state abbreviation (e.g., "IN", "DEMO")}

\item{year}{The year for which to add the matrices}

\item{cohort_type}{The type of cohort (e.g., "EOC", "EOC_CSEM", "EOC_CSEM_CSEM")}

\item{add.matrices.to.which.state}{The state to which to add the matrices}
}
\value{
The updated SGPstateData object with the added matrices
}
\description{
This function adds SGPFlow matrices to the SGPstateData environment.
It retrieves the appropriate matrices based on the state, year, and cohort type,
and assigns them to the SGPstateData object.
}
\note{
This function is not exported and is intended for internal use only.
}
\keyword{internal}
