% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{betaCopula}
\alias{betaCopula}
\title{Generate Beta Copula}
\usage{
betaCopula(number, quantile, multiplier)
}
\arguments{
\item{number}{The number of samples to generate.}

\item{quantile}{The quantile of the beta distribution.}

\item{multiplier}{The multiplier of the beta distribution.}
}
\value{
A vector of samples from the beta copula.
}
\description{
Generate a beta copula with specified quantile and multiplier.
}
\keyword{internal}
