% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{bound.scores}
\alias{bound.scores}
\title{Bound Scores}
\usage{
bound.scores(scale.scores, loss.hoss)
}
\arguments{
\item{scale.scores}{The scores to bound.}

\item{loss.hoss}{The loss hoss.}
}
\value{
The bounded scores.
}
\description{
Bound scores to the loss hoss.
}
\keyword{internal}
