% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{capWords}
\alias{capWords}
\title{Convert All Caps to Mixed Case}
\usage{
capWords(
  x,
  special.words = c("ELA", "I", "II", "III", "IV", "CCSD", "CUSD", "CUD", "USD", "PSD",
    "UD", "ESD", "DCYF", "EMH", "HS", "MS", "ES", "SES", "IEP", "ELL", "MAD", "PARCC",
    "SBAC", "SD", "SWD", "US", "SGP", "SIMEX", "SS", "SAT", "PSAT", "WIDA", "ACCESS",
    "WIDA-ACCESS")
)
}
\arguments{
\item{x}{A character string or vector to be converted to mixed case.}

\item{special.words}{A character vector specifying words or acronyms that should not be converted to mixed case (e.g., "ELA", "I", "III"). Default includes commonly used acronyms in education and assessment contexts.}
}
\value{
A character string in mixed case, with specific formatting applied.
}
\description{
Converts a character string from all uppercase to mixed case, intelligently handling capitalization and preserving specific words or acronyms in uppercase.

This function is particularly useful for cleaning data where names, titles, or other text fields are stored in all caps. It intelligently converts text to mixed case while preserving the format of numbers, punctuation, and specific words or acronyms that should remain in uppercase.
}
\details{
- Words specified in `special.words` remain in uppercase.
- Numbers and their formats (e.g., "2.0") are preserved.
- Handles punctuation like hyphens, apostrophes, and parentheses appropriately.
- Trims leading/trailing spaces and reduces multiple spaces to a single space.

This function can process individual character strings or vectors of strings (using `sapply` for vectorized operations). For factors, it is recommended to apply the function to the levels of the factor.
}
\examples{
\dontrun{
if(interactive()) {
  # Basic examples
  capWords("TEST")             # "Test"
  capWords("ELA TEST1 TEST2")  # "ELA Test1 Test2"
  capWords("O'NEIL")           # "O'Neil"
  capWords("JOHN'S")           # "John's"

  # Processing a character vector
  test.vector <- paste("TEST", 1:10, sep = "")
  sapply(test.vector, capWords)

  # Handling factors by converting levels
  test.factor <- factor(paste("TEST", rep(letters[1:10], each = 50)))
  levels(test.factor) <- sapply(levels(test.factor), capWords)
  levels(test.factor)
}
}

}
\seealso{
\code{\link[base]{toupper}}, \code{\link[base]{tolower}}
}
