% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{convertTime}
\alias{convertTime}
\title{Convert Time Duration to Human-Readable String}
\usage{
convertTime(tmp.time)
}
\arguments{
\item{tmp.time}{A character string in the format "HH:MM:SS" representing a time duration.}
}
\description{
Converts a time duration in the format "HH:MM:SS" to a human-readable string indicating days, hours, minutes, and seconds.

This function takes a time duration in the format "HH:MM:SS" and converts it to a human-readable string that indicates the number of days, hours, minutes, and seconds. It is particularly useful for displaying time durations in a more understandable format.
}
\keyword{internal}
