% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMatrices.R
\name{createMatrices}
\alias{createMatrices}
\title{FUNCTION_TITLE}
\usage{
createMatrices(
  data_for_matrices,
  state,
  matrix.sgp.config,
  super_cohort.sgp.config,
  parallel.config,
  matrix_types = c("single-cohort", "super-cohort")
)
}
\arguments{
\item{data_for_matrices}{PARAM_DESCRIPTION}

\item{state}{PARAM_DESCRIPTION}

\item{matrix.sgp.config}{PARAM_DESCRIPTION}

\item{super_cohort.sgp.config}{PARAM_DESCRIPTION}

\item{parallel.config}{PARAM_DESCRIPTION}

\item{matrix_types}{PARAM_DESCRIPTION, Default: c("single-cohort", "super-cohort")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[createSuperCohortData][SGP::createSuperCohortData], [prepareSGP][SGP::prepareSGP], [baselineSGP][SGP::baselineSGP]
}
