% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createScaleScorePermutations.R
\name{createScaleScorePermutations}
\alias{createScaleScorePermutations}
\title{FUNCTION_TITLE}
\usage{
createScaleScorePermutations(state, sgp.config, permutations = "ALL")
}
\arguments{
\item{state}{PARAM_DESCRIPTION}

\item{sgp.config}{PARAM_DESCRIPTION}

\item{permutations}{PARAM_DESCRIPTION, Default: 'ALL'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\note{
This function is not exported and is intended for internal use only.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[melt][data.table::melt], [data.table][data.table::data.table], [set][data.table::set], [setcolorder][data.table::setcolorder], [rbindlist][data.table::rbindlist]
 [sgpFlowStateData][sgpFlow::sgpFlowStateData]
}
\keyword{internal}
