% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSuperCohortData.R
\name{createSuperCohortData}
\alias{createSuperCohortData}
\title{FUNCTION_TITLE}
\usage{
createSuperCohortData(
  base_data,
  sgp.config,
  supercohort_base_years,
  indicate_cohort = FALSE
)
}
\arguments{
\item{base_data}{PARAM_DESCRIPTION}

\item{sgp.config}{PARAM_DESCRIPTION}

\item{supercohort_base_years}{Subset of years of supplied_base data to use for super-cohort construction}

\item{indicate_cohort}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[data.table][data.table::data.table], [rbindlist][data.table::rbindlist], [setkey][data.table::setkey]
}
