% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{ddcast}
\alias{ddcast}
\title{Data Table dcast}
\usage{
ddcast(tmp.dt, ...)
}
\arguments{
\item{tmp.dt}{The data table to cast.}

\item{...}{Additional arguments to pass to the dcast function.}
}
\value{
A data table in wide format.
}
\description{
Casts a data table to a wide format.
}
\keyword{internal}
