% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLookupTables.R
\name{extractLookupTables}
\alias{extractLookupTables}
\title{FUNCTION_TITLE}
\usage{
extractLookupTables(sgp_object)
}
\arguments{
\item{sgp_object}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[tstrsplit][data.table::tstrsplit], [setcolorder][data.table::setcolorder], [melt.data.table][data.table::melt.data.table], [setattr][data.table::setattr], [setnames][data.table::setnames], [setkey][data.table::setkey]
}
