% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{get.loss.hoss}
\alias{get.loss.hoss}
\title{Get Loss Hoss}
\usage{
get.loss.hoss(state, content_area, grade)
}
\arguments{
\item{state}{The state to get the loss hoss for.}

\item{content_area}{The content area to get the loss hoss for.}

\item{grade}{The grade to get the loss hoss for.}
}
\value{
The loss hoss for the given state, content area, and grade. Returns a numeric vector of length 2 for use with Pmin and Pmax.
}
\description{
Get the loss hoss for a given state, content area, and grade.
}
\keyword{internal}
