% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPercentileTrajectories.R
\name{getPercentileTrajectories}
\alias{getPercentileTrajectories}
\title{FUNCTION_TITLE}
\usage{
getPercentileTrajectories(
  wide_data,
  state,
  sgpFlow.config,
  projection.splineMatrices,
  growth.distribution = NULL,
  csem.perturbation.of.initial.scores = TRUE,
  csem.perturbation.iterations = 100L,
  csem.distribution = "Normal"
)
}
\arguments{
\item{wide_data}{PARAM_DESCRIPTION}

\item{state}{PARAM_DESCRIPTION}

\item{sgpFlow.config}{PARAM_DESCRIPTION}

\item{projection.splineMatrices}{PARAM_DESCRIPTION}

\item{growth.distribution}{PARAM_DESCRIPTION, Default: NULL}

\item{csem.perturbation.of.initial.scores}{PARAM_DESCRIPTION, Default: TRUE}

\item{csem.perturbation.iterations}{PARAM_DESCRIPTION, Default: 100}

\item{csem.distribution}{PARAM_DESCRIPTION, Default: 'Normal'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[copy][data.table::copy], [setorder][data.table::setorder], [data.table][data.table::data.table], [rbindlist][data.table::rbindlist]
 [sgpFlowStateData][sgpFlow::sgpFlowStateData]
 [runif][stats::runif]
 [setv][collapse::setv], [fsubset][collapse::fsubset], [collapv][collapse::collapv], [na_omit][collapse::na_omit], [pivot][collapse::pivot], [qDT][collapse::qDT]
 [bs][splines::bs]
}
