% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSplineMatrices.R
\name{getSplineMatrices}
\alias{getSplineMatrices}
\title{Get Spline Matrices}
\usage{
getSplineMatrices(
  my.matrices,
  my.matrix.content_area.progression,
  my.matrix.grade.progression,
  my.matrix.time.progression,
  my.matrix.time.progression.lags,
  my.exact.grade.progression.sequence = FALSE,
  return.highest.order.matrix = FALSE,
  return.multiple.matrices = FALSE,
  my.matrix.order = NULL,
  my.matrix.highest.order = NULL,
  my.matrix.time.dependency = NULL,
  what.to.return = "MATRICES"
)
}
\arguments{
\item{my.matrices}{A list of spline matrices.}

\item{my.matrix.content_area.progression}{A character vector of content areas for progression.}

\item{my.matrix.grade.progression}{A character vector of grades for progression.}

\item{my.matrix.time.progression}{A character vector of years for progression.}

\item{my.matrix.time.progression.lags}{A numeric vector of year lags for progression.}

\item{my.exact.grade.progression.sequence}{A logical value indicating whether to use the exact grade progression sequence.}

\item{return.highest.order.matrix}{A logical value indicating whether to return the highest order matrix.}

\item{return.multiple.matrices}{A logical value indicating whether to return multiple matrices.}

\item{my.matrix.order}{A numeric value indicating the order of the matrix to return.}

\item{my.matrix.highest.order}{A numeric value indicating the highest order of the matrix to return.}

\item{my.matrix.time.dependency}{A character vector of time dependencies for the matrix.}

\item{what.to.return}{A character value indicating what to return.}
}
\description{
This function is used to get the spline matrices associated with the grade progression and progression sequence provided by the supplied arguments to the function.
}
\keyword{internal}
