% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWideData.R
\name{getWideData}
\alias{getWideData}
\title{Convert Long Data to Wide Format for Cohort Analysis}
\usage{
getWideData(
  long_data,
  sgpFlow.config,
  cohort.end.year = NULL,
  achievement.percentiles.tables
)
}
\arguments{
\item{long_data}{A `data.table` containing student-level data in long format, including variables such as \code{ID}, \code{YEAR}, \code{GRADE}, \code{CONTENT_AREA}, and \code{SCALE_SCORE}.}

\item{sgpFlow.config}{A list containing configuration parameters for student growth percentile (SGP) analysis, including:
\describe{
  \item{\code{year_lags.progression}}{A vector of year lags for cohort progression.}
  \item{\code{grade.progression}}{A vector of grade levels for the cohort.}
  \item{\code{content_area.progression}}{A vector of content areas for the cohort.}
}}

\item{cohort.end.year}{An optional integer specifying the cohort's end year. Defaults to the maximum year present in \code{long_data}.}

\item{achievement.percentiles.tables}{Logical. Indicating whether subset based upon the achievement percentile is performed (99 resulting rows)}
}
\value{
A wide-format `data.table` containing the following columns:
  \describe{
    \item{\code{ID}}{Student IDs.}
    \item{\code{SSX}}{SCALE_SCORE values for grades specified in \code{sgpFlow.config}, where \code{X} corresponds to the grade.}
  }
  The final column is filtered to remove rows with \code{NA} values.
}
\description{
The `getWideData` function takes long-format student data and converts it into a wide-format data table for cohort-based analyses.
It aligns years, grades, and content areas based on the provided `sgpFlow.config` and optionally filters by the specified cohort end year.
}
\details{
This function performs the following steps:
\enumerate{
  \item Generates a lookup table for valid cases based on cohort configuration.
  \item Joins the long-format data with the lookup table to retain only matching records.
  \item Converts the data into a wide format using \code{collapse::pivot}.
  \item Filters rows with \code{NA} in the final column of \code{SCALE_SCORE}.
}
}
