% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{messagesgpFlow}
\alias{messagesgpFlow}
\title{Print Messages to console and log file}
\usage{
messagesgpFlow(tmp.message, domain = NULL, appendLF = TRUE)
}
\arguments{
\item{tmp.message}{The message to print.}
}
\description{
Prints messages to console and a log file.
}
\keyword{internal}
