% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{sgpBeta}
\alias{sgpBeta}
\title{Generate Beta Distribution}
\usage{
sgpBeta(
  n,
  mean.sgp = 50,
  sd.sgp = 15,
  sgp.min.value = NULL,
  sgp.max.value = NULL
)
}
\arguments{
\item{n}{The number of samples to generate.}

\item{mean.sgp}{The mean of the beta distribution.}

\item{sd.sgp}{The standard deviation of the beta distribution.}

\item{sgp.min.value}{The minimum value of the beta distribution.}

\item{sgp.max.value}{The maximum value of the beta distribution.}
}
\value{
A vector of samples from the beta distribution.
}
\description{
Generate a beta distribution with specified mean and standard deviation.
}
\keyword{internal}
