% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbScoresWithCSEM.R
\name{perturbScoresWithCSEM}
\alias{perturbScoresWithCSEM}
\title{Perturb Scores with CSEM}
\usage{
perturbScoresWithCSEM(
  wide_data,
  state,
  sgpFlow.config,
  csem.distribution = "Normal",
  csem.perturbation.iterations = 100L,
  trajectory.type
)
}
\arguments{
\item{wide_data}{A data.table of wide data.}

\item{state}{A character value of the state.}

\item{sgpFlow.config}{A list of configuration parameters for the sgpFlow package.}
}
\description{
This function is used to perturb scale scores with associated scale score conditional standard errors of measurement (CSEM) values.
}
\keyword{internal}
