% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgpFlow-package.R
\docType{package}
\name{sgpFlow-package}
\alias{sgpFlow}
\alias{sgpFlow-package}
\title{sgpFlow: Functions for sgpFlow Analyses Utilizing Data Derived from Student Growth Percentile (SGP) Analyses}
\description{
sgpFlow analyses utilize the SGP framework to construct student learning flows allowing for the simultaneous modeling of student growth and attainment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://centerforassessment.github.io/sgpFlow/}
  \item Report bugs at \url{https://github.com/centerforassessment/sgpFlow/issues/}
}

}
\author{
\strong{Maintainer}: Damian W. Betebenner \email{dbetebenner@nciea.org} (\href{https://orcid.org/0000-0003-0476-5599}{ORCID})

Authors:
\itemize{
  \item Adam R. Van Iwaarden \email{avaniwaarden@nciea.org}
}

}
\keyword{internal}
