% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{sgpFlowTestConfigs}
\alias{sgpFlowTestConfigs}
\title{SGP Flow Test Configurations}
\usage{
sgpFlowTestConfigs()
}
\description{
Configuration definitions for sgpFlow package testing
}
\keyword{internal}
