% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsgpFlow.R
\name{testsgpFlow}
\alias{testsgpFlow}
\title{Test sgpFlow}
\usage{
testsgpFlow(test.number, save.results = TRUE, test.option)
}
\arguments{
\item{test.number}{Test number}

\item{save.results}{Save results to disk from tests (default=TRUE)}

\item{test.option}{Test option}
}
\description{
Test of omnibus sgpFlow package functionalities.
}
