% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{timetakensgpFlow}
\alias{timetakensgpFlow}
\title{Calculate Time Taken for sgpFlow Function}
\usage{
timetakensgpFlow(started.at)
}
\arguments{
\item{started.at}{The starting time of the sgpFlow function.}
}
\value{
A character string in the format "HH:MM:SS" indicating the time taken for the function to execute.
}
\description{
Calculates the time taken for an sgpFlow function to execute.

This function takes the starting time of an sgpFlow function and calculates the time taken for the function to execute. It is particularly useful for timing the execution of sgpFlow functions.
}
\keyword{internal}
