% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{yearIncrement}
\alias{yearIncrement}
\title{Increment Year}
\usage{
yearIncrement(base_year, year_lags)
}
\arguments{
\item{base_year}{The base year to increment.}

\item{year_lags}{The number of years to increment the base year by.}
}
\value{
A character string representing the incremented year.
}
\description{
Increments a year by a specified number of years.
}
\keyword{internal}
