% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLookupTables.R
\name{createLookupTables}
\alias{createLookupTables}
\title{FUNCTION_TITLE}
\usage{
createLookupTables(
  state,
  sgp.config,
  parallel.config,
  matrices = NULL,
  lookup_table_types = c("single-cohort", "super-cohort")
)
}
\arguments{
\item{state}{PARAM_DESCRIPTION}

\item{sgp.config}{PARAM_DESCRIPTION}

\item{parallel.config}{PARAM_DESCRIPTION}

\item{matrices}{PARAM_DESCRIPTION, Default: NULL}

\item{lookup_table_types}{PARAM_DESCRIPTION, Default: c("single-cohort", "super-cohort")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[sgpFlowMatrices][sgpFlowMatrices::sgpFlowMatrices]
 [new][methods::new]
 [abcSGP][SGP::abcSGP]
 [setDTthreads][data.table::setDTthreads], [fwrite][data.table::fwrite]
}
