\name{sgpFlow-package}
\alias{sgpFlow-package}
\docType{package}
\title{Functions for sgpFlow Analyses Utilizing Data Derived from SGP Analyses
}
\description{sgpFlow analyses utilize the SGP framework to construct student learning flows allowing for the simultaneous modeling of student growth and attainment.
}

\details{
\tabular{ll}{
Package: \tab sgpFlow\cr
Type: \tab Package\cr
Version: \tab 0.0-0.85\cr
Date: \tab 2024-11-7\cr
License: \tab MIT\cr
LazyLoad: \tab yes\cr
}
}

\author{Damian W. Betebenner \email{DBetebenner@nciea.org}}

\keyword{package}
