% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgpFlow-package.R
\docType{package}
\name{sgpFlow-package}
\alias{sgpFlow-package}
\title{sgpFlow: Tools for Generating sgpFlow Trajectories}
\description{
The `sgpFlow` package provides tools for generating sgpFlow trajectories based on longitudinal student data. 
It includes utilities for computing growth percentile trajectories, handling cohort-specific analyses, perturbing scores based on 
Conditional Standard Error of Measurement (CSEM), and managing state-specific configurations.

The package is designed for scalability and efficiency when working with large datasets, integrating features such as projection spline matrices, 
growth distribution customization, and flexible trajectory simulation.
}
\details{
\tabular{ll}{
  Package: \tab sgpFlow \cr
  Type: \tab Package \cr
  Version: \tab 0.0-0.992 \cr
  Date: \tab 2025-5-12 \cr
  License: \tab MIT \cr
  LazyLoad: \tab yes \cr
}
 
**Main Features:**
- **Trajectory Generation:** Tools like \code{\link[sgpFlow]{sgpFlowTrajectories}} enable computation of student growth trajectories over multiple iterations.
- **Wide and Long Data Support:** Utilities like \code{\link[sgpFlow]{getWideData}} allow seamless conversion between data formats.
- **State-Specific Configurations:** Embedded configurations (\code{sgpFlowStateData}) facilitate state-specific SGP analyses.
- **Simulation Support:** Integrates CSEM perturbation methods for variability simulation.

**Embedded Data:**
The package includes the \code{sgpFlowStateData} object, which provides state-specific meta-data, including:
- **Achievement Data:** Contains information such as CSEM values required for perturbation.
- **Projection Matrices:** State-specific projection spline matrices for modeling student growth.
}
\examples{
\dontrun{
if(interactive()){
  # Generate SGP trajectories
DEMO_Percentile_Trajectories_Cohort_50_LIST <- sgpFlowTrajectories(
                                           long_data = sgpData::sgpData_LONG,
                                           state = "DEMO",
                                           sgpFlow.config = sgpFlow.config,
                                           growth.distribution = "50",
                                           projection.splineMatrices = projection.splineMatrices)
  # Access the first iteration results
  print(DEMO_Percentile_Trajectories_Cohort_50_LIST[[1]])
}
}
}
\seealso{
\code{\link[sgpFlow]{sgpFlowTrajectories}}, 
 \code{\link[sgpFlow]{getPercentileTrajectories}}, 
 \code{\link[sgpFlow]{getWideData}},
}
\author{
\strong{Maintainer}: Damian W. Betebenner \email{dbetebenner@nciea.org} (\href{https://orcid.org/0000-0003-0476-5599}{ORCID})

Authors:
\itemize{
  \item Adam R. Van Iwaarden \email{avaniwaarden@nciea.org}
}

}
\keyword{package}
