% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgpFlow.R
\name{sgpFlow}
\alias{sgpFlow}
\title{sgpFlow: Create sgpFlow trajectories}
\usage{
sgpFlow(
  sgp_object,
  state,
  sgpFlow.config,
  superCohort.config = NULL,
  cohort.data.type = "SINGLE_COHORT",
  csem.perturbation.of.initial.scores = TRUE,
  csem.perturbation.iterations = 100L,
  iterate.without.csem.perturbation = FALSE,
  achievement.percentiles.tables = TRUE,
  projection.splineMatrices
)
}
\arguments{
\item{sgp_object}{An object of class \code{"SGP"} or \code{"data.table"}. If of class \code{"SGP"}, the `@Data` slot is used; otherwise, the supplied \code{data.table} is used directly.}

\item{state}{A character string indicating the state for which the analysis is conducted. This must match a state entry in \code{sgpFlowStateData}.}

\item{sgpFlow.config}{A configuration list for the sgpFlow analysis, containing details such as grade progression, year lags, and content area progression.}

\item{superCohort.config}{A configuration list specific to super-cohort analysis. If \code{NULL}, super-cohort analysis is skipped.}

\item{cohort.data.type}{A character vector specifying the type of cohort data to analyze. Options include \code{"SUPER_COHORT"}, \code{"COHORT"}, and \code{"ACHIEVEMENT_PERCENTILES"}. Default is \code{"COHORT"}.}

\item{csem.perturbation.of.initial.scores}{Logical. If \code{TRUE}, initial scores are perturbed using conditional standard error measurement (CSEM). Default is \code{TRUE}.}

\item{csem.perturbation.iterations}{Integer. Number of iterations for CSEM perturbation. Default is \code{100L}.}

\item{iterate.without.csem.perturbation}{Logical. If `TRUE`, performs CSEM iterations without perturbing score to derive 100 simulated trajectories from single (non-perturbed) initial score.}

\item{achievement.percentiles.tables}{Logical. Indicating whether subset based upon the achievement percentile is performed (99 resulting rows)}

\item{projection.splineMatrices}{A list of projection spline matrices used for calculating growth trajectories.}
}
\value{
A list containing sgpFlow analysis results for each specified \code{cohort.data.type}.
}
\description{
The `sgpFlow` function facilitates the analysis of sgpFlow trajectories using various cohort data types, such as traditional cohorts, super-cohorts, or achievement percentiles. It offers the flexibility to handle conditional standard error measurement (CSEM) perturbation, growth projection matrices, and other configurations.
}
\details{
The `sgpFlow` function loops over specified \code{cohort.data.type} values, performing SGP trajectory analysis for each type. It supports super-cohort analysis when a valid \code{superCohort.config} is provided. If \code{csem.perturbation.of.initial.scores} is enabled, the function applies random perturbations to initial scores to account for measurement error, with the number of iterations controlled by \code{csem.perturbation.iterations}.

The function requires valid projection spline matrices, provided through \code{projection.splineMatrices}, to perform growth trajectory calculations.
}
\examples{
\dontrun{
# Example usage with an SGP object
results <- sgpFlow(
  sgp_object = sgp_object,
  state = "DEMO",
  sgpFlow.config = sgpFlow.config,
  superCohort.config = superCohort.config,
  cohort.data.type = c("COHORT", "SUPER_COHORT"),
  csem.perturbation.of.initial.scores = TRUE,
  csem.perturbation.iterations = 100L,
  projection.splineMatrices = projection.splineMatrices
)
}

}
