% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgpFlowTrajectories.R
\name{sgpFlowTrajectories}
\alias{sgpFlowTrajectories}
\title{FUNCTION_TITLE}
\usage{
sgpFlowTrajectories(
  long_data,
  state,
  sgpFlow.config,
  cohort_end_year = NULL,
  growth.distribution = NULL,
  csem.perturbation.of.initial.scores = TRUE,
  csem.perturbation.iterations = 100L,
  projection.splineMatrices
)
}
\arguments{
\item{long_data}{PARAM_DESCRIPTION}

\item{state}{PARAM_DESCRIPTION}

\item{sgpFlow.config}{PARAM_DESCRIPTION}

\item{cohort_end_year}{PARAM_DESCRIPTION, Default: NULL}

\item{growth.distribution}{PARAM_DESCRIPTION, Default: NULL}

\item{csem.perturbation.of.initial.scores}{PARAM_DESCRIPTION, Default: TRUE}

\item{csem.perturbation.iterations}{PARAM_DESCRIPTION, Default: 100}

\item{projection.splineMatrices}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
